Internals

LinearCombinations.Hashed — Type

LinearCombinations.Hashed{T}

Hashed{T}(x::T) is a wrapper that stores hash(x) along with x. Computing the hash of such an element returns the stored hash value.

See also LinearCombinations.unhash.

LinearCombinations.unhash — Function

LinearCombinations.unhash(x)

Return the argument unless it is of the form Hashed{T}(y), in which case y is returned.

See also LinearCombinations.Hashed.

LinearCombinations.unval — Function

LinearCombinations.unval(x)

Return c if the argument x is of type Val{c} and x itself otherwise.

This can be used to write type-stable code for the coefftype keyword argument to linear and multilinear functions.

See also @linear, @linear_kw, @multilinear.

Examples

julia> LinearCombinations.unval(Char)
Char

julia> LinearCombinations.unval(Val(Char))
Char

LinearCombinations.Zero — Type

Zero

Zero is a type whose only value Zero() behaves like 0, but allows for simplification at compile time.

LinearCombinations.LinearStyle — Type

LinearCombinations.LinearStyle

The broadcasting style used for linear combinations.

See also @linear_broadcastable.